Left recursion in fewer lines due to new method in CompiledParser.
The left recursion update code is now in CompiledParser#left_recursion_update.
This trims the size of generated grammars significantly.  The subject code runs
when a left-recursive parse is active, and decides what to do with the most
recent parse attempt: continue growing, fall back to previous parse, or fail
altogether.